fix(build): unblock clang/msvc -Werror from lock-refactor commits
Two -Werror breakages landed with the lock-coverage / trigger commits
(CI red since 2026-07-26, all configs):
- abi_dd_ri_test.cpp:342 — unused local `locked` (MSVC C4189 /
  clang -Wunused-variable). Removed; the lock is taken via AdsLockTable.
- session.cpp:1033 — `u16 | (u16 << 8)` promotes to int, then the
  assignment truncates (clang -Wimplicit-int-conversion). Wrapped the
  OR in a static_cast<uint16_t>, matching the idiom used elsewhere.

Build only — does not address the separate 18 abi_dd_trigger_fire_test
failures (AdsAppendRecord/WriteRecord failing under the server lock
try+retry refactor, 044d127f), which are Antonio's/Pritpal's lock work.

+++ src/network/session.cpp
@@ -1029,9 +1029,9 @@ DispatchResult Session::dispatch(const Frame& f) {
             auto open_mode = openads::engine::OpenMode::Shared;
             if (client_open_table_mode_ok_ && f.payload.size() >= 2) {
                 // M12.x extended payload: [u16 LE mode][table_name_bytes]
-                std::uint16_t mode_u16 =
+                std::uint16_t mode_u16 = static_cast<std::uint16_t>(
                     static_cast<std::uint16_t>(f.payload[0]) |
-                    (static_cast<std::uint16_t>(f.payload[1]) << 8);
+                    (static_cast<std::uint16_t>(f.payload[1]) << 8));
                 open_mode = static_cast<openads::engine::OpenMode>(mode_u16);
                 rel.assign(f.payload.begin() + 2, f.payload.end());
             } else {
